Feature #11841
Put the setting and the getting of CMIS properties into a single place:
CmisProperties. By the way, document also through the methods of the
CmisProperties class how the versioning, the PWC and the setting of
creation and update data (user and date) works in CMIS as it can differ
from the way Silverpeas works.

Code the support of the Publication creation (as a CMIS folder) in our
CMIS implementation.