Feature #11841
Add a builder in SimpleAttachment to simplify the construction of such
instances.

Refine the creation of a publication in Silverpeas through CMIS. Now,
the adding of a new CMIS object into a CMIS folder is taken in charge by
the walker dedicated to the type of the CMIS folder and it delegates the
creation of the object to the walker in charge of the CMIS object to
create (and hence of the mapped Silverpeas object).
For example: the creation of a publication into a folder of a Kmelia
instance is done either by the  TreeWalkerForComponentInst instance
(for the root folder) or by the TreeWalkerForNodeDetail instance (for
any other inner folders) and these walkers delegate the creation of
the publication itself to the TreeWalkerForPublicationDetail walker.

Add the attachment of a document to a publication.